+++ zerver/tests/test_invite.py
@@ -1627,15 +1627,18 @@ class InvitationsTestCase(InviteUserBase):
         hamlet = self.example_user("hamlet")
         othello = self.example_user("othello")
 
-        streams = []
+        stream_objects = []
+        stream_ids = []
         for stream_name in ["Denmark", "Scotland"]:
-            streams.append(get_stream(stream_name, user_profile.realm))
+            stream = get_stream(stream_name, user_profile.realm)
+            stream_objects.append(stream)
+            stream_ids.append(stream.id)
 
         invite_expires_in_minutes = 2 * 24 * 60
         do_invite_users(
             user_profile,
             ["TestOne@zulip.com"],
-            streams,
+            stream_objects,
             invite_expires_in_minutes=invite_expires_in_minutes,
         )
 
@@ -1646,7 +1649,7 @@ class InvitationsTestCase(InviteUserBase):
             do_invite_users(
                 user_profile,
                 ["TestTwo@zulip.com"],
-                streams,
+                stream_objects,
                 invite_expires_in_minutes=invite_expires_in_minutes,
             )
             do_create_multiuse_invite_link(
@@ -1664,7 +1667,10 @@ class InvitationsTestCase(InviteUserBase):
         )
 
         do_create_multiuse_invite_link(
-            hamlet, PreregistrationUser.INVITE_AS["MEMBER"], invite_expires_in_minutes
+            hamlet,
+            PreregistrationUser.INVITE_AS["MEMBER"],
+            invite_expires_in_minutes,
+            stream_objects,
         )
 
         result = self.client_get("/json/invites")
@@ -1676,6 +1682,7 @@ class InvitationsTestCase(InviteUserBase):
         self.assertEqual(invites[0]["email"], "TestOne@zulip.com")
         self.assertTrue(invites[1]["is_multiuse"])
         self.assertEqual(invites[1]["invited_by_user_id"], hamlet.id)
+        self.assertEqual(set(invites[1]["stream_ids"]), set(stream_ids))
 
     def test_get_never_expiring_invitations(self) -> None:
         self.login("iago")
@@ -2347,3 +2354,145 @@ class MultiuseInviteTest(ZulipTestCase):
             },
         )
         self.assert_json_error(result, "Invalid invite_as")
+
+    def test_multiuse_invite_edits_with_malformed_parameters(self) -> None:
+        """
+        For a successful edit to a multiuse_invite, it is essential
+        that the following parameters are correct.
+        * stream_ids - IDs of the streams the invite applies to
+        * id - ID of the multiuse_invite
+        """
+        self.user_profile = self.example_user("iago")
+
+        initial_invited_as = PreregistrationUser.INVITE_AS["REALM_ADMIN"]
+        modified_invited_as = PreregistrationUser.INVITE_AS["MODERATOR"]
+        initial_streams = [get_stream("Denmark", self.user_profile.realm)]
+        modified_streams = [
+            get_stream(stream_name, self.user_profile.realm).id
+            for stream_name in ["Denmark", "Verona"]
+        ]
+
+        self.login("iago")
+        invite_expires_in_minutes = 2 * 24 * 60
+        do_create_multiuse_invite_link(
+            self.user_profile,
+            initial_invited_as,
+            invite_expires_in_minutes,
+            initial_streams,
+        )
+        initial_multiuse_object = MultiuseInvite.objects.get()
+
+        # Try to patch the invite with a non-existent stream IDs
+        result = self.client_patch(
+            f"/json/invites/multiuse/{initial_multiuse_object.id}",
+            {
+                "invite_as": modified_invited_as,
+                "stream_ids": [1000],
+            },
+        )
+        self.assert_json_error(result, f"Stream does not exist with id: {1000}.")
+
+        # Try to patch the invite with a non-existent invite ID
+        result = self.client_patch(
+            f"/json/invites/multiuse/{initial_multiuse_object.id + 10}",
+            {
+                "invite_as": modified_invited_as,
+                "stream_ids": modified_streams,
+            },
+        )
+        self.assert_json_error(
+            result,
+            f"Invite does not exist with id: {initial_multiuse_object.id + 10}.",
+        )
+
+    def test_non_previleged_multiuse_invite_edits(self) -> None:
+        """
+        For a successful edit to a multiuse_invite, it is essential
+        that the following situation occurs.
+        * The editing user must be an admin
+        * If the link invites users as owners, the editing user must
+          be an owner themselves.
+        """
+        self.user_profile = self.example_user("iago")
+
+        initial_invited_as = PreregistrationUser.INVITE_AS["REALM_ADMIN"]
+        initial_streams = [get_stream("Denmark", self.user_profile.realm)]
+        modified_streams = [
+            get_stream(stream_name, self.user_profile.realm).id
+            for stream_name in ["Denmark", "Verona"]
+        ]
+
+        self.login("iago")
+        invite_expires_in_minutes = 2 * 24 * 60
+        do_create_multiuse_invite_link(
+            self.user_profile,
+            initial_invited_as,
+            invite_expires_in_minutes,
+            initial_streams,
+        )
+        initial_multiuse_object = MultiuseInvite.objects.get()
+
+        # Try to patch the invite as a non-admin user
+        self.login("hamlet")
+        result = self.client_patch(
+            f"/json/invites/multiuse/{initial_multiuse_object.id}",
+            {
+                "invite_as": PreregistrationUser.INVITE_AS["MEMBER"],
+                "stream_ids": modified_streams,
+            },
+        )
+        self.assert_json_error(result, "Must be an organization administrator")
+
+        # Try to patch the invite as a non-owner while
+        # setting invited_as to "REALM_OWNER"
+        self.login("iago")
+        result = self.client_patch(
+            f"/json/invites/multiuse/{initial_multiuse_object.id}",
+            {
+                "invite_as": PreregistrationUser.INVITE_AS["REALM_OWNER"],
+                "stream_ids": modified_streams,
+            },
+        )
+        self.assert_json_error(result, "Must be an organization owner")
+
+    def test_admin_can_edit_multiuse_invite(self) -> None:
+        """
+        This is a sanity check test that occurs in the best-case scenario,
+        taken from the two tests above.
+        """
+        self.user_profile = self.example_user("iago")
+        self.login("iago")
+
+        initial_invited_as = PreregistrationUser.INVITE_AS["REALM_ADMIN"]
+        modified_invited_as = PreregistrationUser.INVITE_AS["MEMBER"]
+        initial_streams = [get_stream("Denmark", self.user_profile.realm)]
+        modified_streams = [
+            get_stream(stream_name, self.user_profile.realm).id
+            for stream_name in ["Denmark", "Verona"]
+        ]
+        modified_multiuse_object_params = {
+            "invite_as": modified_invited_as,
+            "stream_ids": modified_streams,
+        }
+
+        invite_expires_in_minutes = 2 * 24 * 60
+        do_create_multiuse_invite_link(
+            self.user_profile,
+            initial_invited_as,
+            invite_expires_in_minutes,
+            initial_streams,
+        )
+
+        initial_multiuse_object = MultiuseInvite.objects.get()
+        result = self.client_patch(
+            f"/json/invites/multiuse/{initial_multiuse_object.id}", modified_multiuse_object_params
+        )
+        self.assert_json_success(result)
+
+        # Check if the patch succeeds
+        modified_multiuse_object = MultiuseInvite.objects.get(id=initial_multiuse_object.id)
+        self.assertEqual(modified_multiuse_object.invited_as, modified_invited_as)
+        self.assertEqual(
+            list(modified_multiuse_object.streams.values_list("id", flat=True)),
+            modified_streams,
+        )