+++ api_docs/changelog.md
@@ -20,6 +20,13 @@ format used by the Zulip server that they are interacting with.
 
 ## Changes in Zulip 7.0
 
+**Feature level 171**
+
+* `PATCH /invites/multiuse`: Added support for editing multiuse invites
+  to users as admin with regards to role and associated streams.
+* `GET /invites`: Added the `stream_ids` field to be returned for all
+  multiuse invite objects.
+
 **Feature level 170**
 
 * [`POST /user_topics`](/api/update-user-topic):